requirements.txt
influxdb==5.2.2
python-nest==4.1.0
